Restructure application into different modules
Routes.py has been renamed to run.py so to run the application
run:

    python run.py

Flask app creation and it's config have been moved to an app
packages which requires that the templates and static content
is moved as well. Config variables can now also be set via
environment variables. This is more in line with the "12 factor"
web app: http://12factor.net/

The csv_utils.py have been moved out of the root and into a utils
package as the csv module and other utility functions have been
moved out of run.py (routes.py) into their own utils modules.

This might not be the best way to structure the app, but this will
be better to maintain as time goes by.