Merge handle_csv into the upload csv view
It was weird to call a function from the view which returned a
rendered template with no indication that it should. Further look
at the handle_csv function which returned the template shows that
it only validates the csv and renders the templates (plural because
it also rendered the error templates). It is better to see from
the view itself all possible outputs instead of pushing them into
a different function.

This also fixes a bug in that the os.path.splitext has a leading
dot in the file extension so it would check for '.csv' not only
'csv' when the configuration uses 'csv' (which is better for
configuration).